这把真稳了 （still a little problem of deleting file）

+++ attribute_type.h
@@ -5,17 +5,17 @@
 #include <ostream>
 #include <math.h>
 
-inline int dcmp(double x) {
+inline int dcmp(float x) {
     return fabs(x)<1e-9 ? 0 : x<0 ? -1 : 1;
 }
 struct AttrType {
     int idata;
-    double fdata;
+    float fdata;
     std::string sdata;
     int t;
     AttrType():             t(-1)           {}
     AttrType(int d):        idata(d),t(0)    {}
-    AttrType(double d):     fdata(d),t(1)    {}
+    AttrType(float d):     fdata(d),t(1)    {}
     AttrType(std::string d):     sdata(d),t(2)    {}    
     friend bool operator<(const AttrType &lhs, const AttrType &rhs) {
         int t=lhs.t;
